+++ .ctree.cfg
@@ -1,0 +1,12 @@
+[omp]
+cc = gcc-4.9
+cflags = -std=c99 -O3 -fopenmp -fgcse
+
+[c]
+cc = gcc-4.9
+cflags = -std=c99 -O3 -fgcse
+
+[jit]
+compile_path = ./compiled
+cache = True
+
